+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Added
 
+- Added `queueAnalyticsLoggingReadEnabled`, `queueAnalyticsLoggingWriteEnabled`,
+  and `queueAnalyticsLoggingDeleteEnabled` to `azure_storage_account`.
 - Added `blobAnalyticsLoggingReadEnabled`, `blobAnalyticsLoggingWriteEnabled`,
   and `blobAnalyticsLoggingDeleteEnabled` to `azure_storage_account`.